Require Boost for builds.
This involved install boost-devel and libtool RPMs on RHEL6.  It
also required moving m4 files to a new directory.  Then changes
to configure.ac and Makefile.am could be introduced.

There is nothing *using* Boost at this time, but the configuration
checks are there.  The boost.m4 README says that LTLIBRARIES must
be used for libs that use Boost... ?